tests/wpt/tests/device-bound-session-credentials/refresh-replaces-config.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>DBSC refresh sends back new session config</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="helper.js" type="module"></script>

<script type="module">
  import { expireCookie, waitForCookie, addCookieAndSessionCleanup, configureServer, setupShardedServerState, documentHasCookie } from "./helper.js";

  promise_test(async t => {
    await setupShardedServerState();
    const expectedCookieAndValue1 = "auth_cookie=abcdef0123";
    const expectedCookieAndAttributes1 = `${expectedCookieAndValue1};Domain=${get_host_info().ORIGINAL_HOST};Path=/device-bound-session-credentials`;
    addCookieAndSessionCleanup(t, expectedCookieAndAttributes1);
    const expectedCookieAndValue2 = "other_cookie=ghijkl4567";
    const expectedCookieAndAttributes2 = `${expectedCookieAndValue2};Domain=${get_host_info().ORIGINAL_HOST};Path=/device-bound-session-credentials`;
    addCookieAndSessionCleanup(t, expectedCookieAndAttributes2);

    // Prompt starting a session, and wait until registration completes.
    const login_response = await fetch('login.py');
    assert_equals(login_response.status, 200);
    assert_true(await waitForCookie(expectedCookieAndValue1));

    // Confirm that a request has the cookie set.
    const auth_response = await fetch('verify_authenticated.py');
    assert_equals(auth_response.status, 200);
    // Confirm that a request does not have alternate cookie set.
    const alternate_auth_response = await fetch('verify_authenticated.py', {
      method: 'POST',
      body: expectedCookieAndValue2
    });
    assert_equals(alternate_auth_response.status, 401);

    // Configure server to change the cookie in the session config on next refresh.
    configureServer({ cookieNameAndValue: expectedCookieAndValue2 });

    // Expire the first cookie and send a request, which triggers the refresh with the new session config.
    expireCookie(expectedCookieAndAttributes1);
    assert_false(documentHasCookie(expectedCookieAndValue1));
    const auth_response_after_expiry1 = await fetch('verify_authenticated.py');
    assert_equals(auth_response_after_expiry1.status, 401);
    assert_false(documentHasCookie(expectedCookieAndValue1));

    // Confirm the alternate cookie is set and included in requests.
    assert_true(documentHasCookie(expectedCookieAndValue2));
    const alternate_auth_response_after_expiry1 = await fetch('verify_authenticated.py', {
      method: 'POST',
      body: expectedCookieAndValue2
    });
    assert_equals(alternate_auth_response_after_expiry1.status, 200);

    // Expire the second cookie. Confirm the second cookie is refreshed, and not the first.
    expireCookie(expectedCookieAndAttributes2);
    assert_false(documentHasCookie(expectedCookieAndValue2));
    const alternate_auth_response_after_expiry2 = await fetch('verify_authenticated.py', {
      method: 'POST',
      body: expectedCookieAndValue2
    });
    assert_equals(alternate_auth_response_after_expiry2.status, 200);
    assert_true(documentHasCookie(expectedCookieAndValue2));
    assert_false(documentHasCookie(expectedCookieAndValue1));
}, "Refresh can replace session config");
</script>
